主体完成

+++ atom-core/src/router/tree/art/node/direct_node.rs
@@ -73,7 +73,7 @@ impl<N> NodeTrait<N> for DirectNode<N> {
     }
 
     #[inline]
-    fn width(&self) -> u16 {
+    fn width(&self) -> usize {
         256
     }
 }
